requirements.txt
langchain
langgraph
python-dotenv
langchain-aws
langchain-community
streamlit
boto3